+++ ssvm/tests/unittests_data_structures.py
@@ -77,6 +77,71 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey1")
         self.assertEqual(1923, candidates.get_n_cand(spectrum))
 
+    def test_get_constant_ms_score(self):
+        # ----------
+        # SPECTRUM 1
+        # ----------
+        spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "AU88550178"})
+        candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
+
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="CONST_MS_SCORE")
+        self.assertEqual(3934, len(scores))
+        self.assertTrue(np.all(np.array(scores) == 1.0))
+
+    def test_get_multiple_ms2_scores(self):
+        # ----------
+        # SPECTRUM 1
+        # ----------
+
+        # - MetFrag has not scored that spectrum
+        # - SIRIUS scores are available
+
+        spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "LQB6372613"})
+        candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
+
+        for s2w in [0, 0.32, 0.5, 0.72, 1]:
+            scores = candidates.get_ms_scores(
+                spectrum, ms_scorer=["sirius__sd__correct_mf", "metfrag__norm_after_merge"],
+                ms_scorer_weights=[1 - s2w, s2w]
+            )
+
+            self.assertEqual(118, len(scores))
+            self.assertTrue(np.all(np.array(scores) > 0))
+            self.assertEqual(1.0, np.max(scores))
+
+            _scores_sirius = candidates.get_ms_scores(
+                spectrum, ms_scorer="sirius__sd__correct_mf", return_as_ndarray=True
+            )
+            np.testing.assert_equal(((1 - s2w) * _scores_sirius + s2w), scores)
+
+        # ----------
+        # SPECTRUM 2
+        # ----------
+        spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "BML0302194"})
+        candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
+
+        for s2w in [0, 0.32, 0.5, 0.72, 1]:
+            scores = candidates.get_ms_scores(
+                spectrum, ms_scorer=["sirius__sd__correct_mf", "metfrag__norm_after_merge"],
+                ms_scorer_weights=[1 - s2w, s2w]
+            )
+
+            self.assertEqual(1271, len(scores))
+            self.assertTrue(np.all(np.array(scores) > 0))
+            self.assertEqual(1.0, np.max(scores))
+
+            _scores_sirius = candidates.get_ms_scores(
+                spectrum, ms_scorer="sirius__sd__correct_mf", return_as_ndarray=True
+            )
+            _scores_metfrag = candidates.get_ms_scores(
+                spectrum, ms_scorer="metfrag__norm_after_merge", return_as_ndarray=True
+            )
+
+            _scores_comb = ((1 - s2w) * _scores_sirius + s2w * _scores_metfrag)
+            _scores_comb /= np.max(_scores_comb)
+
+            np.testing.assert_equal(_scores_comb, scores)
+
     def test_get_ms2_scores__normalized(self):
         # ----------
         # SPECTRUM 1
@@ -85,13 +150,13 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf")
         self.assertEqual(3934, len(scores))
         self.assertTrue(np.all(np.array(scores) > 0))
         self.assertEqual(1.0, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
         self.assertEqual(3934, len(scores))
         self.assertEqual(0.0627934092603395, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
@@ -103,13 +168,13 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf")
         self.assertEqual(2308, len(scores))
         self.assertTrue(np.all(np.array(scores) > 0))
         self.assertEqual(1.0, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
         self.assertEqual(2308, len(scores))
         self.assertEqual(0.11534319471384565, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
@@ -121,13 +186,13 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf")
         self.assertEqual(118, len(scores))
         self.assertTrue(np.all(np.array(scores) > 0))
         self.assertEqual(1.0, np.max(scores))
 
         # # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
         self.assertEqual(118, len(scores))
         self.assertEqual(1.0, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
@@ -140,13 +205,13 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
         self.assertEqual(3934, len(scores))
         self.assertEqual(-516.7701408961747, np.min(scores))
         self.assertEqual(-60.65141796101968, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
         self.assertEqual(3934, len(scores))
         self.assertEqual(59.08480943917096, np.min(scores))
         self.assertEqual(940.939664451204, np.max(scores))
@@ -158,31 +223,31 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
         self.assertEqual(2308, len(scores))
         self.assertEqual(-616.7732484911674, np.min(scores))
         self.assertEqual(-138.34323578753632, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
         self.assertEqual(2308, len(scores))
         self.assertEqual(111.97300932463722, np.min(scores))
         self.assertEqual(970.7812377005032, np.max(scores))
 
         # ----------
-        # SPECTRUM 2
+        # SPECTRUM 3
         # ----------
         spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "LQB6372613"})
         candidates = CandSQLiteDB_Massbank(MASSBANK_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is SIRIUS
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="sirius__sd__correct_mf", scale_scores_to_range=False)
         self.assertEqual(118, len(scores))
         self.assertEqual(-307.381287577836, np.min(scores))
         self.assertEqual(-147.8959133560828, np.max(scores))
 
-        # # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
+        # MS2 Scorer is MetFrag
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge", scale_scores_to_range=False)
         self.assertEqual(118, len(scores))
         self.assertEqual(1e-6, np.min(scores))
         self.assertEqual(1e-6, np.max(scores))
@@ -572,6 +637,17 @@ class TestMassbankCandidateSQLiteDB(unittest.TestCase):
         candidates._ensure_molecule_identifier_is_available("inchikey")
         candidates._ensure_molecule_identifier_is_available("cid")
 
+    def test_ensure_ms2_scorer_is_available(self):
+        candidates = CandSQLiteDB_Massbank(db_fn=MASSBANK_DB_FN)
+
+        with self.assertRaises(ValueError):
+            candidates._ensure_ms2scorer_is_available("super_metfrag")
+            candidates._ensure_ms2scorer_is_available(None)
+            candidates._ensure_ms2scorer_is_available("")
+
+        candidates._ensure_ms2scorer_is_available("sirius__sd__correct_mf")
+        candidates._ensure_ms2scorer_is_available("metfrag__norm_after_merge")
+
 
 class TestCandidateSQLiteDB(unittest.TestCase):
     @staticmethod
@@ -617,13 +693,13 @@ class TestCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Bach2020(BACH2020_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is IOKR
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="IOKR__696a17f3")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="IOKR__696a17f3")
         self.assertEqual(2233, len(scores))
         self.assertEqual(0.00006426196626211542, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="MetFrag_2.4.5__8afe4a14")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="MetFrag_2.4.5__8afe4a14")
         self.assertEqual(2233, len(scores))
         np.testing.assert_allclose(0.000028105936908001407, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
@@ -635,13 +711,13 @@ class TestCandidateSQLiteDB(unittest.TestCase):
         candidates = CandSQLiteDB_Bach2020(BACH2020_DB_FN, molecule_identifier="inchikey")
 
         # MS2 Scorer is IOKR
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="IOKR__696a17f3")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="IOKR__696a17f3")
         self.assertEqual(16, len(scores))
         self.assertEqual(0.7441697188705315, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
 
         # MS2 Scorer is MetFrag
-        scores = candidates.get_ms2_scores(spectrum, ms2scorer="MetFrag_2.4.5__8afe4a14")
+        scores = candidates.get_ms_scores(spectrum, ms_scorer="MetFrag_2.4.5__8afe4a14")
         self.assertEqual(16, len(scores))
         self.assertEqual(0.18165470462229025, np.min(scores))
         self.assertEqual(1.0, np.max(scores))
@@ -986,7 +1062,7 @@ class TestCandidateSQLiteDB(unittest.TestCase):
         # Do not enforce the ground truth structure to be in the candidate set
         candidates = CandSQLiteDB_Bach2020(db_fn=BACH2020_DB_FN, molecule_identifier="inchikey")
 
-        scores = candidates.get_ms2_scores(spectrum, "MetFrag_2.4.5__8afe4a14", return_dataframe=True)
+        scores = candidates.get_ms_scores(spectrum, "MetFrag_2.4.5__8afe4a14", return_dataframe=True)
         fps = candidates.get_molecule_features(spectrum, "iokr_fps__positive", return_dataframe=True)
         labspace = candidates.get_labelspace(spectrum)
 
@@ -1041,6 +1117,59 @@ class TestCandidateSQLiteDB(unittest.TestCase):
 
 
 class TestRandomSubsetCandidateSQLiteDB_Massbank(unittest.TestCase):
+    def test_get_ms_scores(self):
+        # ----------
+        # SPECTRUM 1
+        # ----------
+        spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "BS40569952",
+                                                         "molecule_id": "CKEXCBVNKRHAMX"})
+
+        # Do not enforce the ground truth structure to be in the candidate set
+        candidates = RandomSubsetCandSQLiteDB_Massbank(
+            number_of_candidates=102, db_fn=MASSBANK_DB_FN, molecule_identifier="inchikey1", include_correct_candidate=False)
+
+        # MetFrag Scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
+        self.assertEqual(102, len(df_scores))
+
+        # Constant candidate scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="CONST_MS_SCORE")
+        self.assertEqual(102, len(df_scores))
+
+        # ----------
+        # SPECTRUM 2
+        # ----------
+        spectrum = Spectrum(np.array([]), np.array([]), {"spectrum_id": "KW13980462",
+                                                         "molecule_id": "YPLYFEUBZLLLIY-UHFFFAOYSA-N"})
+
+        # Enforce the ground truth structure to be in the candidate set
+        candidates = RandomSubsetCandSQLiteDB_Massbank(
+            number_of_candidates=12, db_fn=MASSBANK_DB_FN, molecule_identifier="inchikey", include_correct_candidate=True)
+
+        # MetFrag Scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
+        self.assertEqual(12, len(df_scores))
+
+        # Constant candidate scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="CONST_MS_SCORE")
+        self.assertEqual(12, len(df_scores))
+
+        # --------------------
+
+        # Enforce the ground truth structure to be in the candidate set
+        candidates = RandomSubsetCandSQLiteDB_Massbank(
+            number_of_candidates=100000, db_fn=MASSBANK_DB_FN, molecule_identifier="inchikey",
+            include_correct_candidate=True
+        )
+
+        # MetFrag Scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="metfrag__norm_after_merge")
+        self.assertEqual(3834, len(df_scores))
+
+        # Constant candidate scores
+        df_scores = candidates.get_ms_scores(spectrum, ms_scorer="CONST_MS_SCORE")
+        self.assertEqual(3834, len(df_scores))
+
     def test_get_labelspace(self):
         # ----------
         # SPECTRUM 1
@@ -1091,7 +1220,7 @@ class TestRandomSubsetCandidateSQLiteDB_Massbank(unittest.TestCase):
         candidates = RandomSubsetCandSQLiteDB_Massbank(
             number_of_candidates=102, db_fn=MASSBANK_DB_FN, molecule_identifier="inchikey", include_correct_candidate=True)
 
-        scores = candidates.get_ms2_scores(spectrum, "sirius__sd__correct_mf", return_dataframe=True)
+        scores = candidates.get_ms_scores(spectrum, "sirius__sd__correct_mf", return_dataframe=True)
         fps = candidates.get_molecule_features(spectrum, "sirius_fps", return_dataframe=True)
         labspace = candidates.get_labelspace(spectrum)
 
@@ -1213,7 +1342,7 @@ class TestRandomSubsetCandidateSQLiteDB(unittest.TestCase):
         candidates = RandomSubsetCandSQLiteDB_Bach2020(
             number_of_candidates=102, db_fn=BACH2020_DB_FN, molecule_identifier="inchikey", include_correct_candidate=True)
 
-        scores = candidates.get_ms2_scores(spectrum, "MetFrag_2.4.5__8afe4a14", return_dataframe=True)
+        scores = candidates.get_ms_scores(spectrum, "MetFrag_2.4.5__8afe4a14", return_dataframe=True)
         fps = candidates.get_molecule_features(spectrum, "iokr_fps__positive", return_dataframe=True)
         labspace = candidates.get_labelspace(spectrum)
 
@@ -1587,12 +1716,12 @@ class TestBugsAndWiredStuff(unittest.TestCase):
         test_sequences_metfrag = SequenceSample(
             [self.spectra[idx] for idx in test], [self.labels[idx] for idx in test],
             CandSQLiteDB_Bach2020(db_fn=BACH2020_DB_FN, molecule_identifier="inchikey1", init_with_open_db_conn=True),
-            N=50, L_min=30, L_max=50, random_state=19, ms2scorer="MetFrag_2.4.5__8afe4a14")
+            N=50, L_min=30, L_max=50, random_state=19, ms_scorer="MetFrag_2.4.5__8afe4a14")
 
         test_sequences_iokr = SequenceSample(
             [self.spectra[idx] for idx in test], [self.labels[idx] for idx in test],
             CandSQLiteDB_Bach2020(db_fn=BACH2020_DB_FN, molecule_identifier="inchikey1", init_with_open_db_conn=True),
-            N=50, L_min=30, L_max=50, random_state=19, ms2scorer="IOKR__696a17f3")
+            N=50, L_min=30, L_max=50, random_state=19, ms_scorer="IOKR__696a17f3")
 
         for idx in range(len(test_sequences_metfrag)):
             self.assertListEqual(test_sequences_metfrag[idx].get_labelspace(),